% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayRepulsive.R
\name{BayRepulsive_unknown}
\alias{BayRepulsive_unknown}
\title{BayRepulsive_unknown is the NMF function used when the number of subclones is unknown}
\source{
BayRepulsive: A Bayesian Repulsive Deconvolution Model for Inferring Latent Biologic Structure
}
\usage{
BayRepulsive_unknown(Datause, K_min, K_max, Nobs, Nfeature, Niter = 100, epsilon = 0.0001,tau = 100, seed = 1 )
}
\arguments{
\item{Datause}{The observed data matrix. Each row is a sample.}

\item{K_min}{The minimum number of subclones}

\item{K_max}{The Maximum number of subclones}

\item{Nobs}{The number of samples, i.e., the number of rows of the Datause}

\item{Nfeature}{The number of features, i.e., the number of columns of the Datause}

\item{Niter}{The number of maximum iterations}

\item{epsilon}{break if the L2 distance of the two estiamted proportion matrix in row is less than epsilon}

\item{tau}{The hyperparameter for DPP, a large number is prefered, default value is 100}

\item{seed}{The random seed, default as 1}
}
\value{
W, the estiamted signature matrix.

Z, and the estiamted number of subclones.

K, the estimated number of subclones.
}
\description{
Takes in the observed data matrix, the range of number of subclones, the number of features and samples
Gives the estimation of the NMF, including the estimated number of subclones.
}
\details{
Given an observed matrix, whose rows are mixed samples of unknown number of subclones,
we give an estimation of number of subclones, and do the NMF.

This function will create a bunch of globel variables, named "Datause", "Nobs", "Nfeature", "sigma0",
"mu0", "K", "Theta", "W.star", "Z.star",  "W_temp", "sigma.square", "data.now", "Z.now", "i".
Thus, users should avoid these variable names when using BayRepulsive_unknown, if they don't want the variables to be overwritten.
Especially "i", which is commonly used in loops.
}
