% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE11058.R
\docType{data}
\name{GSE11058}
\alias{GSE11058}
\title{GSE11058}
\format{a data frame with 12 variables (observations of pure samples).
Four kinds of immune cell lines: Jurkat, IM-9, Raji, THP-1.
Each kind has three duplications.}
\usage{
data(GSE11058)
}
\description{
This is the pure cell line expression data from GSE11058
}
\examples{
# input the data
data(GSE11058)
# get the name of samples
str(GSE11058)
}
\keyword{datasets}
