% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inhouse.R
\docType{data}
\name{Inhouse}
\alias{Inhouse}
\title{Inhouse}
\format{a data frame with three variables(expression profiles of three kinds of T cells).
Three kinds of T cell lines:
naive CD4+ T cell, naive CD8+ T cell, and activated CD4+ T cell in tumor sample}
\usage{
data(Inhouse)
}
\description{
This is the pure cell line expression data from GSE11058
}
\examples{
# input the data
data(Inhouse)
# get the name of samples
str(Inhouse)
}
\keyword{datasets}
